bnb/requirement.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
Werkzeug
email_validator